+++ src/VaultSharp/V1/Commons/BackendConfig.cs
@@ -35,31 +35,5 @@ namespace VaultSharp.Core
         /// </value>
         [JsonProperty("max_lease_ttl")]
         public int MaximumLeaseTtl { get; set; }
-
-        // raja todo: this sounds non-generic.
-
-        /// <summary>
-        /// Gets or sets the token type.
-        /// </summary>
-        /// <value>
-        /// The token type.
-        /// </value>
-        [JsonProperty("token_type")]
-        public string TokenType { get; set; }
-
-        /// <summary>
-        /// Gets or sets the description.
-        /// </summary>
-        [JsonProperty("description")]
-        public string Description { get; set; }
-
-        /// <summary>
-        /// Gets or sets the plugin name.
-        /// </summary>
-        /// <value>
-        /// The plugin name.
-        /// </value>
-        [JsonProperty("plugin_name")]
-        public string PluginName { get; set; }
     }
 }