+++ src/VaultSharp/V1/Commons/BackendListingVisibility.cs
@@ -1,0 +1,15 @@
+﻿using Newtonsoft.Json;
+using Newtonsoft.Json.Converters;
+
+namespace VaultSharp.Core
+{
+    /// <summary>
+    /// Represents the BackendListingVisibility
+    /// </summary>
+    [JsonConverter(typeof(StringEnumConverter))]
+    public enum BackendListingVisibility
+    {
+        hidden,
+        unauth
+    }
+}