+++ src/VaultSharp/V1/SystemBackend/ISystemBackend.cs
@@ -128,7 +128,7 @@ namespace VaultSharp.V1.SystemBackend
         /// <returns>
         /// A task
         /// </returns>
-        Task ConfigureAuthBackendAsync(string path, BackendConfig backendConfig);
+        Task ConfigureAuthBackendAsync(string path, NewBackendConfig backendConfig);
 
         /// <summary>
         /// Gets the capabilities of the token on the given path.
@@ -508,7 +508,7 @@ namespace VaultSharp.V1.SystemBackend
         /// <returns>
         /// A task
         /// </returns>
-        Task ConfigureSecretBackendAsync(string path, BackendConfig backendConfig);
+        Task ConfigureSecretBackendAsync(string path, NewBackendConfig backendConfig);
 
         /// <summary>
         /// Gets all the available policy names in the system.